Add denominator adjusted SLO
Add recording rules for storing how the current number of events
compares to the average of the preceding week. This allows an SLO to
burn faster or slower depending on whether there are a lot or few events
in the denominator.

This will make sure that SLO's burn slower at night when there are few
events.