feat: error budget exhausted plugin
This adds a plugin that creates a new alert for total error budget
exhaustion. It can be useful to initiate a change freeze, SLO review,
retrospective, or whatever an organization does when an error budget is
totally exhausted.

Everything is customizable and it'll pull in existing labels in line
with the other SLOs.

I also updated codeowners with slok, myself and cxdy.